requirements.txt
langchain-community
langchain-openai
langchain-core
langchain
openai
python-dotenv
faiss-cpu
boto3
PyPDF